requirements.txt
colorama==0.4.6
Django==4.2.11
django-bootstrap-v5==1.0.11
django-bootstrap5==23.4
django-debug-toolbar==4.3.0
psycopg2==2.9.9